Added network/router lsa origination

+++ VirtualRouter/src/protocols/ospf/ospfv2/database/ExternalLsaV2.hpp
@@ -51,7 +51,7 @@ struct ExternalLsaV2
         return true;
     }
 
-    static constexpr size_t size()
+    static constexpr uint16_t size()
     {
         return 16;
     }
