+++ VirtualRouter/src/protocols/ospf/ospfv2/database/RouterLsaV2.hpp
@@ -89,9 +89,9 @@ struct RouterLsaV2
         return true;
     }
 
-    inline size_t size() const
+    inline uint16_t size() const
     {
-        return 4 + (4 * links.size());
+        return 4 + static_cast<uint16_t>(4 * links.size());
     }
 
     void appendChecksum(ChecksumFletcher& check) const
