+++ VirtualRouter/src/protocols/ospf/ospfv2/database/SummaryNetworkLsa.hpp
@@ -36,7 +36,7 @@ struct SummaryNetworkLsa
         return true;
     }
 
-    static constexpr size_t size()
+    static constexpr uint16_t size()
     {
         return 8;
     }
