+++ VirtualRouter/src/protocols/ospf/ospfv2/database/SummaryRouterLsa.hpp
@@ -38,7 +38,7 @@ struct SummaryRouterLsa
         return true;
     }
 
-    static constexpr size_t size()
+    static constexpr uint16_t size()
     {
         return 8;
     }
